Reversão de movimentação de fontes.
