+++ fontes/avaliador-sintatico/README.md
