+++ fontes/avaliador-sintatico/dialetos/index.ts
