+++ fontes/avaliador-sintatico/index.ts
