+++ fontes/avaliador-sintatico/retornos/index.ts
