+++ fontes/avaliador-sintatico/retornos/retorno-primario.ts
@@ -10,8 +10,8 @@ import {
     Super,
     Variavel,
     Vetor,
-} from '../../../construtos';
-import { Importar } from '../../../declaracoes';
+} from '../../construtos';
+import { Importar } from '../../declaracoes';
 
 export type RetornoPrimario =
     | Super
