+++ fontes/avaliador-sintatico/traducao/README.md
