+++ fontes/avaliador-sintatico/traducao/avaliador-sintatico-javascript.ts
