+++ fontes/index.ts
@@ -1,4 +1,4 @@
-export * from './analisador-semantico/avaliador-sintatico';
+export * from './avaliador-sintatico';
 export * from './construtos';
 export * from './declaracoes';
 export { PontoParada, cyrb53 } from './depuracao';
