+++ fontes/interfaces/retornos/retorno-avaliador-sintatico.ts
@@ -1,4 +1,4 @@
-import { ErroAvaliadorSintatico } from '../../analisador-semantico/avaliador-sintatico/erro-avaliador-sintatico';
+import { ErroAvaliadorSintatico } from '../../avaliador-sintatico/erro-avaliador-sintatico';
 
 export interface RetornoAvaliadorSintatico<T> {
     declaracoes: T[];
