+++ fontes/tradutores/tradutor-calango.ts
@@ -1,4 +1,4 @@
-import { AvaliadorSintaticoPortugolIpt } from '../analisador-semantico/avaliador-sintatico/dialetos';
+import { AvaliadorSintaticoPortugolIpt } from '../avaliador-sintatico/dialetos';
 import { FormatacaoEscrita, Literal } from '../construtos';
 import { LexadorCalango } from '../lexador/dialetos';
